score to uint16

+++ inc/PhysicsEngine.h
@@ -31,6 +31,6 @@ typedef struct {
 } frame_t;
 
 //function declarations
-void updatePhysics(ball_t * ball_p, uint8_t * activeBalls_p, uint32_t * striker0_p, uint32_t * striker1_p, uint8_t * lives_p, uint8_t * score_p);
+void updatePhysics(ball_t * ball_p, uint8_t * activeBalls_p, uint32_t * striker0_p, uint32_t * striker1_p, uint8_t * lives_p, uint16_t * score_p);
 
 #endif /* _PHYSICSENGINE_H_ */
